refactor(screens): update app background
